chore(ci): add PWM HAL unit tests, enforce pylint and mypy
CI:
- tests.yml: add c-unit-tests job (ubuntu-latest, cmake build + run
  test_pwm_hal); run alongside existing Python test matrix
- examples.yml: add Run PWM HAL tests step to cpp-tests job

tests.yml:
- Add 'fix/**' to push branch triggers so CI fires on push (not only PR)
- c-unit-tests job: build all targets, run via ctest --output-on-failure
  (was running ./build/test_pwm_hal directly, silently skipping test_arduino_bridge)

examples.yml:
- Add 'fix/**' to push branch triggers
- Add lgpio-pwm-fade and lgpio-pwm-servo to build matrix (exclude Windows,
  same as lgpio-blink — ARM cross-compile setup required)
- cpp-tests job: remove duplicate PWM HAL test step (tests.yml owns it)

Pylint / mypy (enforce — remove continue-on-error):
- .pylintrc: created with targeted disables for PlatformIO-specific
  false positives (C0415, R0903, C0103, C0302, W0223, R0914, R0913,
  R0917, R0911, C0413, W0603)
- platform_config.py, platform.py, platform_test_uploader.py: fix all
  pylint and mypy violations; expand mypy scope to all four .py files
- tests/test_platform.py: fix pylint violations
- .pre-commit-config.yaml: align black version and line-length

Examples:
- lgpio-pwm-fade, lgpio-pwm-servo: replace atoi() CLI argument parsing
  with strtol() + range validation for clear invalid-argument errors

Further Corrections:
- Add libmsgpack-cxx-dev to c-unit-tests CI apt install (CRITICAL: cmake
  configure was aborting before any test ran)
- Add platformio>=6.1,<7 to requirements-dev.txt; remove duplicate pip
  install from lint and coverage-report jobs (supply chain hardening)
- Fix docs/TESTING.md: 20 unit tests -> 29 unit tests
- Add C unit test build/run instructions to CONTRIBUTING.md
- Expand tests/ directory tree in AGENTS.md (adds test_pwm_hal.c and
  stub file listing)
- Replace hardcoded /sys/class/pwm/ literals in sysfs stub with
  PWM_SYSFS_BASE macro to stay in sync with production header
- Add max_len == 0 guard in pwm_sysfs_read() before size_t underflow
- Switch duty-cycle intermediate arithmetic from float to double for
  correct precision at low PWM frequencies (< 8 kHz)

+++ framework-lgpio/pwm-hal.c
@@ -360,7 +360,7 @@ int pwm_write(int pin, float duty_cycle_percent) {
 
     // Calculate duty cycle in nanoseconds
     uint64_t period_ns = 1000000000ULL / state->frequency_hz;
-    uint64_t duty_cycle_ns = (uint64_t)(period_ns * (duty_cycle_percent / 100.0f));
+    uint64_t duty_cycle_ns = (uint64_t)((double)period_ns * (duty_cycle_percent / 100.0));
 
     // Set duty cycle
     char path[MAX_PATH_LEN];
@@ -441,7 +441,7 @@ int pwm_set_frequency(int pin, uint32_t freq_hz) {
     }
 
     // Update duty cycle to maintain percentage
-    uint64_t duty_cycle_ns = (uint64_t)(period_ns * (state->duty_cycle_percent / 100.0f));
+    uint64_t duty_cycle_ns = (uint64_t)((double)period_ns * (state->duty_cycle_percent / 100.0));
     snprintf(path, sizeof(path), "%s/pwmchip%d/pwm%d/duty_cycle",
              PWM_SYSFS_BASE, state->pwm_chip, state->pwm_channel);
     snprintf(value, sizeof(value), "%llu", (unsigned long long)duty_cycle_ns);
@@ -544,7 +544,7 @@ int pwm_get_status(int pin, pwm_status_t *status) {
     status->duty_cycle_percent = state->duty_cycle_percent;
     status->polarity = state->polarity;
     status->period_ns = 1000000000ULL / state->frequency_hz;
-    status->duty_cycle_ns = (uint64_t)(status->period_ns * (state->duty_cycle_percent / 100.0f));
+    status->duty_cycle_ns = (uint64_t)((double)status->period_ns * (state->duty_cycle_percent / 100.0));
 
     return PWM_SUCCESS;
 }
